requirements.txt
anyio==4.4.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
azure-core==1.30.2
azure-storage-blob==12.22.0
boto3==1.34.155
botocore==1.34.155
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssselect2==0.7.0
Django==5.1.1
django-jazzmin==3.0.0
django-storages==1.14.4
django_paystack==1.0.0
et-xmlfile==1.1.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
isodate==0.6.1
jmespath==1.0.1
lxml==5.3.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
paystackapi==2.1.1
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypaystack2==2.0.3
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
